Update&Create
Summary of changes:

1. Create the struct & initialize it [push_swaw.h lines: 18 to 31; stack.c lines: 13 to 47]
2. Reorganize the project structure

Details:

First, I created the new_node function to allocate and initialize a new node.

Then, I implemented the push function to insert nodes into the stack. This function updates the next and prev pointers to properly link the new node: the new node's next points to the former head, its prev is set to NULL since it is now the top of the stack, the old head's prev is updated to point back to the new node, and finally, the stack pointer is updated to this new node.

Pending tasks:
I still need to verify/test this implementation.