Update
Summary of changes

1. Corrected the sort_3 algorithm
[
  - algorithims.c lines: 15 to 40
]

2. Order 4 parameters

[
  - turc_algorithim_utils.c new file
  - algorithims lines:  42 to 50
]

The problem with the sort 3 was that it doesn't cover all the
posibilities. The solution was searching for the biggest number and
depending of it's position it moves with the ra rra or sa.

With the sort_4 the solution was, searching for the lowest number of the
stack then push it into the stack B moving it first on the top of stack
a. When the lowest number was into the stack B the program sort the 3
numbers left and finally the lowest number was pushed back into the
stack a